Exchange animation
Was created the logic to simulate a 'shuffle' of the participants an
then show the participant name to give a gift.

Changes applieds:

 - exchange_controller: Was added one line in the index controller, was
   selected only the name of all record in the participant table and was
   saved in a @particpants instance variable.
 - application.js: Was added just one line that requires the raffle.js
   file.
 - raffle.js: Contians the logic to show the participants, and some
   styles for the animation that founds in the _exchange.scss file.
   Before execute this functions, is executed a setTimeOut to wait for a
   100 miliseconds, this to wait that all the DOM is loaded, but this
   'issue can' be solved with differents ways. Also is imported a calls
   called 'Animations' that founds in the animations.js.
 - animations.js: Contains a class called Animations with a method
   called raffle, that receives a value, this value is the element that
   contains the name of the particpant.

The animation to raffle a participant only lasts 10 seconds, after taht
the print_gift_to_participant function is called, the opacity is set to
0, and after one second is set the display property to none, this is to wait
that the element disapear slowly and not roughly and after disapear, the
name with the particpant appears slowly.

The index fiel contains the classes and containers for the view, also
has a script tag where are defined the main variables that after will be
used for the JS files.