Modifying the raffle animation
Was deleted the variables and properties to set the left, right, top and
bottom values due to in the .scss files, a display flex is used, and
this properties can be used with a display flex.

The idea of the raffle method in the animations class is set the random
time for the animation duration and the animation delay. The entire
animation was created in the _exchange.scss file.