Current user validation
Was added a validation to validate if the gift_to field form the current
user is nil or not, if true, then the animation is executed, else the
animation won't be executed and the directly the name of the participant
to give a gift is showed.

Also was changed the animation duration from 2s to 1.3s.